refactor: reduce component to default card element
