+++ scripts/postCss.js
@@ -15,11 +15,30 @@ fs.readFile('src/style.css', (err, css) => {
         ':root': '*'
       }
     }))
-    .process(css, { from: 'src/style.css', to: 'src/style.css' })
+    .process(css, { from: 'src/style-hero.css', to: 'src/style-hero.css' })
     .then(result => {
-      fs.writeFile('src/style.css', result.css, () => true)
+      fs.writeFile('src/style-hero.css', result.css, () => true)
       if ( result.map ) {
-        fs.writeFile('src/style.map', result.map, () => true)
+        fs.writeFile('src/style-hero.map', result.map, () => true)
+      }
+    })
+})
+
+fs.readFile('src/style-hero.css', (err, css) => {
+  postcss([autoprefixer, postcssCustomProperties, comments])
+    .use(comments({
+      remove: function(comment) { return comment[0] == "@"; }
+    }))
+    .use(removeRules({
+      rulesToRemove: {
+        ':root': '*'
+      }
+    }))
+    .process(css, { from: 'src/style-hero.css', to: 'src/style-hero.css' })
+    .then(result => {
+      fs.writeFile('src/style-hero.css', result.css, () => true)
+      if ( result.map ) {
+        fs.writeFile('src/style-hero.map', result.map, () => true)
       }
     })
 })
